--- renfxx/set_renf_xalloc.cpp
@@ -1,18 +1,0 @@
-/*
-    Copyright (C) 2018 Vincent Delecroix
-
-    This file is part of e-antic
-
-    e-antic is free software: you can redistribute it and/or modify it under
-    the terms of the GNU Lesser General Public License (LGPL) as published
-    by the Free Software Foundation; either version 3.0 of the License, or
-    (at your option) any later version.  See <http://www.gnu.org/licenses/>.
-*/
-
-#include <e-antic/renfxx.h>
-
-int eantic::renf_class::xalloc()
-{
-    static int xa = std::ios_base::xalloc();
-    return xa;
-}
